feat: remove standalone app, fully integrate redirectors into main app
Remove the entire `redirector_app/` standalone application. All redirector
functionality is now served by the main cyberx-event-mgmt backend.

Changes:
- Port standalone detail.html and list.html (with all PR #20 features) to
  frontend/templates/ with {{ url_prefix }} for /admin path prefix
- Migrate log_buffer.py to backend/app/services/ and add /admin/logs page
- Add Error Logs link to main app sidebar (admin.manage permission)
- Add SecurityHeadersMiddleware (X-Frame-Options, X-Content-Type-Options,
  Referrer-Policy) — previously only in standalone, now in main app
- Install memory log handler in main app lifespan for log viewer
- Delete: redirector_app/, standalone Dockerfile, INTEGRATION.md

The main app now has all features that were standalone-only:
- Prerequisites accordion with auto-fix
- nginx setup banner with auto-fix
- Port conflict checking with row highlighting
- Deployed column, OS info display, dark mode fixes
- Error log viewer at /admin/logs
