+++ frontend/templates/pages/admin/logs.html
