--- redirector_app/.env.example
@@ -1,54 +1,0 @@
-# ============================================================================
-# CyberX Redirector Manager — Environment Configuration
-# Copy this file to .env and fill in all required values.
-# NEVER commit .env to version control.
-# ============================================================================
-
-# ── SSH Key Encryption (REQUIRED) ──────────────────────────────────────────
-# Fernet symmetric encryption key for SSH private keys stored in the database.
-# Generate: python -c "from cryptography.fernet import Fernet; print(Fernet.generate_key().decode())"
-FERNET_KEY=
-
-# ── API Authentication (REQUIRED) ──────────────────────────────────────────
-# Secret key for X-API-Key header authentication (automation / API clients).
-# Generate: python -c "import secrets; print(secrets.token_urlsafe(32))"
-API_KEY=
-
-# ── Web UI Admin Credentials (REQUIRED for web UI) ─────────────────────────
-# Username for the web UI login page.
-ADMIN_USERNAME=admin
-
-# Bcrypt hash of the admin password.
-# Generate: python -c "import bcrypt; print(bcrypt.hashpw(b'yourpassword', bcrypt.gensalt(12)).decode())"
-ADMIN_PASSWORD_HASH=
-
-# Email address shown in the sidebar footer (cosmetic only).
-ADMIN_EMAIL=admin@localhost
-
-# ── JWT Session Secret (REQUIRED for web UI) ───────────────────────────────
-# HS256 signing secret for JWT session cookies.
-# Generate: python -c "import secrets; print(secrets.token_urlsafe(32))"
-SECRET_KEY=
-
-# Optional: separate CSRF signing secret. Defaults to SECRET_KEY if unset.
-# Generate: python -c "import secrets; print(secrets.token_urlsafe(32))"
-CSRF_SECRET_KEY=
-
-# JWT session lifetime in minutes (default: 480 = 8 hours)
-SESSION_EXPIRY_MINUTES=480
-
-# ── Database (PostgreSQL required) ────────────────────────────────────────
-# SQLAlchemy async database URL. Start PostgreSQL with: docker compose up -d postgres
-DATABASE_URL=postgresql+asyncpg://cyberx:changeme@localhost:5432/cyberx_events
-
-# ── Application ────────────────────────────────────────────────────────────
-# Set to "production" to enable Secure cookie flag on the session cookie.
-# In development (without HTTPS), leave this unset or set to "development".
-APP_ENV=development
-
-# Allowed CORS origins as JSON array. Empty array disables CORS (recommended).
-# Example: CORS_ORIGINS=["https://yourdomain.com"]
-CORS_ORIGINS=[]
-
-# Application version shown in the dashboard footer (optional).
-APP_VERSION=1.0.0
