--- redirector_app/alembic/env.py
@@ -1,55 +1,0 @@
-"""Alembic environment for the standalone redirector app."""
-import asyncio
-import os
-from logging.config import fileConfig
-
-from sqlalchemy.ext.asyncio import create_async_engine
-from alembic import context
-
-from dotenv import load_dotenv
-load_dotenv()
-
-config = context.config
-if config.config_file_name:
-    fileConfig(config.config_file_name)
-
-# Import models so their metadata is registered
-from redirector_app.models import Base  # noqa: E402
-target_metadata = Base.metadata
-
-DATABASE_URL = os.environ.get(
-    "DATABASE_URL", "postgresql+asyncpg://cyberx:changeme@localhost:5432/cyberx_events"
-)
-
-
-def run_migrations_offline() -> None:
-    context.configure(
-        url=DATABASE_URL,
-        target_metadata=target_metadata,
-        literal_binds=True,
-        dialect_opts={"paramstyle": "named"},
-    )
-    with context.begin_transaction():
-        context.run_migrations()
-
-
-def do_run_migrations(connection):
-    context.configure(
-        connection=connection,
-        target_metadata=target_metadata,
-    )
-    with context.begin_transaction():
-        context.run_migrations()
-
-
-async def run_migrations_online() -> None:
-    connectable = create_async_engine(DATABASE_URL)
-    async with connectable.connect() as connection:
-        await connection.run_sync(do_run_migrations)
-    await connectable.dispose()
-
-
-if context.is_offline_mode():
-    run_migrations_offline()
-else:
-    asyncio.run(run_migrations_online())
