--- redirector_app/requirements.txt
@@ -1,36 +1,0 @@
-# Standalone Redirector Manager requirements
-# Run: pip install -r requirements.txt
-
-# FastAPI and server
-fastapi==0.115.6
-uvicorn[standard]==0.34.0
-python-multipart==0.0.20
-jinja2>=3.1.4
-
-# Database (PostgreSQL required)
-sqlalchemy==2.0.36
-greenlet>=3.1.0         # required by SQLAlchemy async engine
-asyncpg==0.30.0         # PostgreSQL async driver
-alembic==1.14.0
-
-# Encryption (SSH key storage)
-cryptography==44.0.0
-
-# SSH/SFTP
-paramiko==4.0.0
-
-# Validation
-pydantic==2.10.5
-pydantic-settings==2.7.1
-email-validator==2.2.0
-
-# Web UI auth
-bcrypt==4.2.0
-python-jose[cryptography]==3.3.0
-itsdangerous==2.2.0
-
-# Static files serving
-aiofiles==24.1.0
-
-# Environment
-python-dotenv==1.0.1
