--- redirector_app/web/templates/pages/auth/login.html
@@ -1,89 +1,0 @@
-{% extends "layouts/auth.html" %}
-
-{% block title %}Login{% endblock %}
-
-{% block content %}
-<div class="container">
-    <div class="row justify-content-center">
-        <div class="col-lg-5">
-            <div class="card cyber-card border-0 mt-5">
-                <div class="card-body">
-                    <div class="text-center pt-4 pb-2">
-                        <img src="/static/img/cyberx-star-logo.svg" alt="CyberX" style="width: 56px; height: 56px;" class="mb-3 cyber-logo">
-                        <h3 class="cyber-title">CyberX</h3>
-                        <div class="cyber-subtitle mt-1">Redirector Manager</div>
-                    </div>
-
-                    <div id="loginError" class="alert alert-danger alert-dismissible fade show mt-3 d-none" role="alert">
-                        <span id="loginErrorMsg"></span>
-                        <button type="button" class="btn-close" data-bs-dismiss="alert"></button>
-                    </div>
-
-                    <form id="loginForm" class="mt-4">
-                        <div class="mb-3">
-                            <label class="small mb-1" for="username">Username</label>
-                            <input class="form-control" id="username" name="username" type="text"
-                                   placeholder="Enter username" required autofocus autocomplete="username" />
-                        </div>
-                        <div class="mb-3">
-                            <label class="small mb-1" for="password">Password</label>
-                            <input class="form-control" id="password" name="password" type="password"
-                                   placeholder="Enter password" required autocomplete="current-password" />
-                        </div>
-                        <div class="d-flex align-items-center justify-content-end mt-4 mb-0">
-                            <button type="submit" class="btn btn-cyber" id="loginBtn">Login</button>
-                        </div>
-                    </form>
-                </div>
-                <div class="card-footer text-center py-3">
-                    <div class="small text-muted">
-                        CyberX Redirector Manager
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock %}
-
-{% block extra_js %}
-<script>
-document.getElementById('loginForm').addEventListener('submit', async function(e) {
-    e.preventDefault();
-
-    const btn = document.getElementById('loginBtn');
-    const errEl = document.getElementById('loginError');
-    const errMsg = document.getElementById('loginErrorMsg');
-
-    errEl.classList.add('d-none');
-    btn.disabled = true;
-    btn.textContent = 'Authenticating...';
-
-    const username = document.getElementById('username').value.trim();
-    const password = document.getElementById('password').value;
-
-    try {
-        const response = await csrfFetch('/api/web/login', {
-            method: 'POST',
-            body: JSON.stringify({ username, password }),
-            headers: { 'Content-Type': 'application/json' }
-        });
-
-        if (response.ok) {
-            window.location.href = '/redirectors';
-        } else {
-            const data = await response.json().catch(() => ({}));
-            errMsg.textContent = data.detail || 'Login failed. Please check your credentials.';
-            errEl.classList.remove('d-none');
-            btn.disabled = false;
-            btn.textContent = 'Login';
-        }
-    } catch (error) {
-        errMsg.textContent = 'An error occurred. Please try again.';
-        errEl.classList.remove('d-none');
-        btn.disabled = false;
-        btn.textContent = 'Login';
-    }
-});
-</script>
-{% endblock %}
